requirements.txt
numpy>=1.24.0
torch>=2.4.1
openslide-python>=1.3.1
openslide-bin>=4.0.0.8
timm==1.0.8
scikit-image>=0.24.0
opencv-python-headless>=4.10.0.84
pandas>=1.5
tqdm>=4.66.5
loguru>=0.7.2
typer
shapely>=2.1.0
joblib>=1.4.2
xformers>=0.0.29.post1
huggingface-hub>=0.33.4
matplotlib>=3.7.0

